requirements.txt
pyyaml
requests
pandas 
openpyxl
smartapi-python
logzero
websocket-client
nsepython
